Renamed file and automated output
#1 
Automated the output as a pdf. Didn't close the issue because I was wondering if you want the output as an html because it would keep the interactivity and improve the image quality (doesn't require an internet connection). The only qualm I have is that I can't find a way to download an html file through github.