#4 Edit SSB budget share formula and fix typo in SOC function
Edited the SSB budget share formula to be a function of SSB expenditure in dollars rather than SSB consumption in ounces.

Fixed a typo in the compute_SOC function that accidentally called the compute_v_sc function when the compute_v_cc function was intended.

The output files change slightly as a result of these changes.